Fix bug introduced in last merge

+++ backend/src/service/userService.go
@@ -35,7 +35,7 @@ func GetUsers(filter model.User) *[]model.User {
 
 // UpdateUser Update specific user using id param in URL
 func UpdateUser(updatedUser model.User, id int) error {
-	updatedUser.ID = uint8(id)
+	updatedUser.ID = uint(id)
 
 	output := storage.DB.Save(&updatedUser)
